Refactor: level-calculator.testのディレクトリ移行
